fix types in IncidentsTable on MS
