+++ services/common/src/tests/mocks/dataMocks.tsx
@@ -1965,6 +1965,7 @@ export const INCIDENT: IMineIncident = {
   status_code: "CLD",
   mine_name: "Test Mine",
   categories: [{ mine_incident_category_code: "ENV" }],
+  incident_timezone: "Canada/Pacific",
 };
 
 export const INCIDENTS = {
